Move methods and files around
This moves most methods to their final locations, at least for now. The
end goal is to reorganize the code from atool so that it can be used in
other scripts easily (chpass, approve, sorry, etc.).

I've just been cut/pasting methods. So the imports are really screwed
up, and you shouldn't expect any of this code to work. I'll be working
to further clean up the code and fix the imports, then remove the
redundant code from atool and use alib instead.