Import and reorganize account creation/filtering
This imports a lot of code from create, which was written collectively
by OCF staffers, but especially:

- Nader Morshed <morshed@ocf.berkeley.edu> (primary author)
- Kenny Do <kedo@ocf.berkeley.edu> (account filtering)
- Dara Adib <daradib@ocf.berkeley.edu> (realname checking, minor fixes)

Much of the code has been reorganized or slightly rewritten in
preparation of switching to single-account-transaction account creation
(rather than batch processing of several accounts at once).

Since we also no longer use the 'mid' stage of account creation (where
accounts have been approved, but wait 24 hours to be created), that code
is also removed.

Currently, this imports most of the account filtering, but not all of
the account creation. The main things still remaining are:

* LDAP entry creation
* Kerberos principal creation

There are some smaller things still remaining as well:

* Getting max UID

It also hasn't been tested at all.