Fix tests use of mock.called
Although the new mock did break a lot of things, this isn't one of them.

mock.assert_called() is not actually a thing. (There is
mock.assert_called_with() and mock.assert_called_once_with(), though).

The new mock raises an AttributeError when this happens, presumably
because this is a common mistake.