Drop cached_property because of pip bug
We can't install ocflib on the wheezy servers anymore because of a bug
in pip that prevents packages from being installed that have undecodable
bytes (via the current locale, which in Puppet is "C").

This was fixed in https://github.com/pydanny/cached-property/pull/20 but
no release has been made yet.

This can be reverted once a release is made, or when we're fully on
jessie (which installs cached-property via a Debian package).