misc.shell: by default, only color if stdout is a tty
This is a breaking change but I think it's probably one that will
improve behavior everywhere we use it...

If not, setting `tty_only=False` in the calls is the workaround. But I
can't think of an existing case where we'd want that.

This change means that if you dump text to a file, or otherwise execute
it without a proper TTY (e.g. via commands from ocfweb), you won't get
ANSI color codes embedded in the text.