Update lab hours through Fall 2017
These are definitely subject to change, but these are the days we
definitely won't be open. Closing early will probably still be a thing
that happens too, because the lab won't be very busy then anyway and
opstaff probably want to leave for break.

Also fix some years I put in previous and didn't change from copying
and pasting... lol