Don't leave a +3 file around in `make autoversion`
This unwanted file was getting created because dch invokes the
given editor roughly like:

    $VISUAL +linenum debian/changelog

Real editors interpret +linenum as a command to jump to the given
line number, but touch interprets it as another file to create.
In order to remedy this, we supply a message in the dch args,
which prevents dch from opening up an editor.