Fix domain-related test
Also see https://github.com/ocf/ocfweb/pull/571 where this needed to be
fixed in ocfweb. Ultimately I think not reaching out to DNS and mocking
the response would be a better practice so we're not dependent on the
return values of the university's DNS (or even whether it's available or
not).